requirements.txt
djangorestframework
Django==2.2.3
numpy==1.16.4
pandas==0.25.0
psycopg2-binary==2.8.3
python-dateutil==2.8.0
pytz==2019.1
six==1.12.0
sqlparse==0.3.0
xlrd==1.2.0
